+++ clients/all.d.ts
@@ -172,3 +172,4 @@ export import Backup = require('./backup');
 export import WorkLink = require('./worklink');
 export import Textract = require('./textract');
 export import ManagedBlockchain = require('./managedblockchain');
+export import MediaPackageVod = require('./mediapackagevod');